refactor: useDurationOpen 훅을 공용 훅으로 분리
